requirements.txt
 mediapipe
 opencv-python
 numpy
 numpy
 Pillow
